confile: make reading config file thread-safe
When multiple threads read the same config file (e.g. the global default
config) the file pointer seems to be modified by the various threads. There's
an obvious solution to this problem: m{un}map().
